refactor(summary): change monetary summary fields from BigDecimal to Double

+++ src/main/java/com/carlos/expensetracker/dto/response/ExpenseSummaryResponse.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 public record ExpenseSummaryResponse(
         BigDecimal totalAmount,
         Long totalCount,
-        BigDecimal averageAmount,
+        Double averageAmount,
         BigDecimal minAmount,
         BigDecimal maxAmount,
         LocalDate  periodStart,
